Drop custom mapper implementation
Note: We had developed this custom mapper due to
https://github.com/kubernetes-sigs/controller-runtime/issues/490,
however, the problem reported there got already fixed 3y ago, see
https://github.com/kubernetes-sigs/controller-runtime/issues/490#issuecomment-2554169323
for more information.
