+++ pkg/controllerutils/mapper/mapper.go
@@ -6,7 +6,12 @@ package mapper
 
 import (
 	"context"
+	"fmt"
 
+	"k8s.io/apimachinery/pkg/api/meta"
+	"k8s.io/apimachinery/pkg/runtime"
+	"k8s.io/apimachinery/pkg/types"
+	utilruntime "k8s.io/apimachinery/pkg/util/runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/handler"
 	"sigs.k8s.io/controller-runtime/pkg/predicate"
@@ -34,4 +39,31 @@ func ClusterToObjectMapper(reader client.Reader, newObjListFunc func() client.Ob
 			return predicateutils.EvalGeneric(o, predicates...)
 		})
 	}
+}
+
+// ObjectListToRequests adds a reconcile.Request for each object in the provided list.
+func ObjectListToRequests(list client.ObjectList, predicates ...func(client.Object) bool) []reconcile.Request {
+	var requests []reconcile.Request
+
+	utilruntime.HandleError(meta.EachListItem(list, func(object runtime.Object) error {
+		obj, ok := object.(client.Object)
+		if !ok {
+			return fmt.Errorf("cannot convert object of type %T to client.Object", object)
+		}
+
+		for _, predicate := range predicates {
+			if !predicate(obj) {
+				return nil
+			}
+		}
+
+		requests = append(requests, reconcile.Request{NamespacedName: types.NamespacedName{
+			Name:      obj.GetName(),
+			Namespace: obj.GetNamespace(),
+		}})
+
+		return nil
+	}))
+
+	return requests
 }
