+++ pkg/controllerutils/mapper/mapper_test.go
@@ -11,6 +11,7 @@ import (
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 	"go.uber.org/mock/gomock"
+	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/types"
 	"sigs.k8s.io/controller-runtime/pkg/client"
@@ -106,6 +107,47 @@ var _ = Describe("Controller Mapper", func() {
 
 		It("should find no objects because the passed object is no cluster", func() {
 			Expect(mapper(ctx, infra)).To(BeEmpty())
+		})
+	})
+
+	Describe("#ObjectListToRequests", func() {
+		list := &corev1.SecretList{
+			Items: []corev1.Secret{
+				{ObjectMeta: metav1.ObjectMeta{Name: "secret1", Namespace: "namespace1"}},
+				{ObjectMeta: metav1.ObjectMeta{Name: "secret1", Namespace: "namespace2"}},
+				{ObjectMeta: metav1.ObjectMeta{Name: "secret2", Namespace: "namespace2"}},
+				{ObjectMeta: metav1.ObjectMeta{Name: "secret1", Namespace: "namespace3"}},
+				{ObjectMeta: metav1.ObjectMeta{Name: "secret1", Namespace: "namespace4"}},
+				{ObjectMeta: metav1.ObjectMeta{Name: "secret1", Namespace: "namespace5"}},
+				{ObjectMeta: metav1.ObjectMeta{Name: "secret1", Namespace: "namespace6"}},
+			},
+		}
+
+		It("should return the correct requests w/p predicates", func() {
+			Expect(ObjectListToRequests(list)).To(ConsistOf(
+				reconcile.Request{NamespacedName: types.NamespacedName{Name: "secret1", Namespace: "namespace1"}},
+				reconcile.Request{NamespacedName: types.NamespacedName{Name: "secret1", Namespace: "namespace2"}},
+				reconcile.Request{NamespacedName: types.NamespacedName{Name: "secret2", Namespace: "namespace2"}},
+				reconcile.Request{NamespacedName: types.NamespacedName{Name: "secret1", Namespace: "namespace3"}},
+				reconcile.Request{NamespacedName: types.NamespacedName{Name: "secret1", Namespace: "namespace4"}},
+				reconcile.Request{NamespacedName: types.NamespacedName{Name: "secret1", Namespace: "namespace5"}},
+				reconcile.Request{NamespacedName: types.NamespacedName{Name: "secret1", Namespace: "namespace6"}},
+			))
+		})
+
+		It("should return the correct requests w/ predicates", func() {
+			var (
+				predicate1 = func(o client.Object) bool { return o.GetNamespace() != "namespace3" }
+				predicate2 = func(o client.Object) bool { return o.GetNamespace() != "namespace5" }
+			)
+
+			Expect(ObjectListToRequests(list, predicate1, predicate2)).To(ConsistOf(
+				reconcile.Request{NamespacedName: types.NamespacedName{Name: "secret1", Namespace: "namespace1"}},
+				reconcile.Request{NamespacedName: types.NamespacedName{Name: "secret1", Namespace: "namespace2"}},
+				reconcile.Request{NamespacedName: types.NamespacedName{Name: "secret2", Namespace: "namespace2"}},
+				reconcile.Request{NamespacedName: types.NamespacedName{Name: "secret1", Namespace: "namespace4"}},
+				reconcile.Request{NamespacedName: types.NamespacedName{Name: "secret1", Namespace: "namespace6"}},
+			))
 		})
 	})
 })
