+++ skaffold-operator.yaml
@@ -366,7 +366,6 @@ build:
             - pkg/client/kubernetes/cache
             - pkg/controller/tokenrequestor
             - pkg/controllerutils
-            - pkg/controllerutils/mapper
             - pkg/controllerutils/predicate
             - pkg/controllerutils/reconciler
             - pkg/controllerutils/routes
